refactor(mcp): split scraper and knowledge-base features into SRP modules
