+++ surfsense_mcp/src/surfsense_mcp/features/scrapers/platforms/__init__.py
@@ -1,0 +1,1 @@
+"""One module per scraper platform; each exposes register(mcp, client, context)."""
